refactor(frontend): landing and hero section moved to pages folder ss-53
