+++ apps/frontend/src/pages/libs/components/hero-section/hero-section.tsx
